+++ python/utils/__init__.py
@@ -5,5 +5,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from .pyqt5patcher import PyQt5Patcher
+from .qtpyimporter import QtPyImporter
 
-__all__ = ('PyQt5Patcher',)
+
+__all__ = ('PyQt5Patcher', 'QtPyImporter')